Refactored Tests and Models
The order model was failing because the user and product data hadn't been entered when it ran.

Refactored the tests to be in a single file (not be practice) to get around this and resolve the issue.

I've renamed the model spec files to remove the word "Spec" so that they aren't included in the test