Handler Tests Working
Stumbled onto a big revelation that the IDs autoincrement when you delete them, they don't reset back to 0/1, instead they continue. The reason it was looking for ID 2 wasn't because it was duplicating the create but rather I had created and deleted a resource with ID 1 and the next in line was 2.

I've since refactored the models, handlers and tests to return better information such as deleting returning the row deleted and using a parameter rather than a body with the ID