Updated Migrations
Updated the migrations used for the project (as per the project rubric).

There's no way to set the order in which they deploy however, you can use the date value at the beginning of the filename to set the order.

I flipped the number around for the first migration I created with the order they needed to be deployed in